+++ tests/testthat/test_peaks_mspct.R
@@ -121,19 +121,19 @@ test_that("source_spct", {
   my.spct <- white_led.source_spct
 
   wls.spct <- wls_at_target(my.spct)
-  expect_equal(wls.spct[["w.length"]], c(540.94, 660.78))
+  expect_equal(wls.spct[["w.length"]], c(540.94, 661.23))
   expect_equal(nrow(wls.spct), 2)
   expect_equal(names(wls.spct), c("w.length", "s.e.irrad"))
   expect_is(wls.spct, "source_spct")
 
   wls.spct <- wls_at_target(my.spct, idfactor = TRUE)
-  expect_equal(wls.spct[["w.length"]], c(540.94, 660.78))
+  expect_equal(wls.spct[["w.length"]], c(540.94, 661.23))
   expect_equal(nrow(wls.spct), 2)
   expect_equal(names(wls.spct), c("w.length", "s.e.irrad", "target.idx"))
   expect_is(wls.spct, "source_spct")
 
   wls.spct <- wls_at_target(my.spct, idfactor = "TARGET")
-  expect_equal(wls.spct[["w.length"]], c(540.94, 660.78))
+  expect_equal(wls.spct[["w.length"]], c(540.94, 661.23))
   expect_equal(nrow(wls.spct), 2)
   expect_equal(names(wls.spct), c("w.length", "s.e.irrad", "TARGET"))
   expect_is(wls.spct, "source_spct")
@@ -155,7 +155,7 @@ test_that("source_spct", {
   expect_equal(wls_at_target(my.spct, target = "0.5 max"), wls.spct)
 
   wls.spct <- wls_at_target(my.spct, target = "half.range")
-  expect_equal(wls.spct[["w.length"]], c(540.94, 660.78))
+  expect_equal(wls.spct[["w.length"]], c(540.94, 661.23))
   expect_equal(nrow(wls.spct), 2)
   expect_equal(names(wls.spct), c("w.length", "s.e.irrad"))
   expect_is(wls.spct, "source_spct")